Fix for #5895 - Changed deprecated numpy.float to float (#5897)
* Removed numpy.float (deprecated) and coverted to float

* Removed numpy.float (deprecated) and coverted to float

+++ rdkit/ML/AnalyzeComposite.py
@@ -65,11 +65,11 @@ def ProcessIt(composites, nToConsider=3, verbose=0):
           levels = TreeUtils.CollectLabelLevels(model, {}, 0, nToConsider)
           TreeUtils.CollectDescriptorNames(model, descNames, 0, nToConsider)
           for descId in levels.keys():
-            v = res.get(descId, numpy.zeros(nToConsider, numpy.float))
+            v = res.get(descId, numpy.zeros(nToConsider, float))
             v[levels[descId]] += 1. / nModels
             res[descId] = v
       for k in res:
-        v = globalRes.get(k, numpy.zeros(nToConsider, numpy.float))
+        v = globalRes.get(k, numpy.zeros(nToConsider, float))
         v += res[k] / nComposites
         globalRes[k] = v
       if verbose > 0:
@@ -113,12 +113,12 @@ def ErrorStats(conn, where, enrich=1):
   if not nPts:
     sys.stderr.write('no runs found\n')
     return None
-  overall = numpy.zeros(nPts, numpy.float)
-  overallEnrich = numpy.zeros(nPts, numpy.float)
+  overall = numpy.zeros(nPts, float)
+  overallEnrich = numpy.zeros(nPts, float)
   oCorConf = 0.0
   oInCorConf = 0.0
-  holdout = numpy.zeros(nPts, numpy.float)
-  holdoutEnrich = numpy.zeros(nPts, numpy.float)
+  holdout = numpy.zeros(nPts, float)
+  holdoutEnrich = numpy.zeros(nPts, float)
   hCorConf = 0.0
   hInCorConf = 0.0
   overallMatrix = None
